Move the runner to a dedicated file
