+++ setup.cfg
@@ -36,6 +36,7 @@ exclude =
     __init__.py
     .tox
     venv
+    dist
 ; max-complexity = 14
 max-line-length = 160
 doctests = True
